+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-05-19
+
 ### Added
 
 - Added `--no-<metric name>` CLI options to disable selected metrics.